Quick benchmarking showed `parent_unittitles_ssm` to be taking up a huge proportion of the indexing time. This refactor speeds things up some.

It has larger effects with more heavily-nested documents.

`bx rake arclight:index FILE=../data/ead/umich-bhl/umich-bhl-2014150.xml` (7.3MB)
* Master (as of 20190920 5pm): 950 seconds
* This branch: 550 seconds (52% of master branch time)

`bx rake arclight:index FILE=../data/ead/InU-ewv/VAD1751.xml` (11MB)
* Master: 243 seconds
* This branch:  179 seconds (74% of master branch time)

(umich-bhl-9677.xml is the file that took @mejackreed 6,000 seconds as per https://code4lib.slack.com/archives/C43V2KVTM/p1568145216166500, so I'm not sure what I'm doing differently -- this is just on my desktop Mac)

I'm 99% sure the refactor is producing identical data (checked by indexing with both methods against a few files) but if someone else could take a look that'd be great.