@mejackreed 2019-09-20
For better context, that 6000 seconds was on an AWS ec2 medium instance, with Solr likely not optimized.